importer/requirements.txt
weaviate-client>=4.10.0
